+++ src/fenceMark.ts
@@ -1,0 +1,15 @@
+export const FenceTag = {
+    START: 'fence:start',
+    END: 'fence:end',
+    SPLIT: 'fence',
+};
+
+export const FenceShortTag = {
+    START: '>>>',
+    END: '<<<',
+    SPLIT: '---',
+};
+
+export function markString(src: string) {
+    return `<!-- ${src} -->`;
+}