+++ src/insertFence.ts
@@ -1,0 +1,4 @@
+export async function insertFence() {
+    console.log("method is empty.");
+
+}